+++ src/extensions.jl
@@ -12,9 +12,9 @@ Base.convert(::Type{LoadStatus}, x::Integer) = LoadStatus(x)
 Base.convert(::Type{LoadModels}, x::Integer) = LoadModels(x)
 Base.convert(::Type{LineUnits}, x::Integer) = LineUnits(x)
 Base.convert(::Type{GeneratorStatus}, x::GeneratorStatus) = GeneratorStatus(x)
-Base.convert(::Type{BatchOperation}, x::BatchOperation) = BatchOperation(x)
 Base.convert(::Type{StorageStates}, x::StorageStates) = StorageStates(x)
 Base.convert(::Type{DSSMessageType}, x::DSSMessageType) = DSSMessageType(x)
+Base.convert(::Type{DSSSaveFlags}, x::DSSSaveFlags) = DSSSaveFlags(x)
 
 # TODO: see https://github.com/dss-extensions/OpenDSSDirect.jl/issues/76
 # Base.convert(::Type{SolverOptions}, x::SolverOptions) = SolverOptions(x)
@@ -77,10 +77,6 @@ $( join(map(x -> "- `" * String(x[1]) * "`: " * string(x[2]), CEnum.name_value_p
 """
 Lib.GeneratorStatus
 @doc """
-$( join(map(x -> "- `" * String(x[1]) * "`: " * string(x[2]), CEnum.name_value_pairs(Lib.BatchOperation)), "\n") )
-"""
-Lib.BatchOperation
-@doc """
 $( join(map(x -> "- `" * String(x[1]) * "`: " * string(x[2]), CEnum.name_value_pairs(Lib.StorageStates)), "\n") )
 """
 Lib.StorageStates
@@ -88,3 +84,7 @@ Lib.StorageStates
 $( join(map(x -> "- `" * String(x[1]) * "`: " * string(x[2]), CEnum.name_value_pairs(Lib.DSSMessageType)), "\n") )
 """
 Lib.DSSMessageType
+@doc """
+$( join(map(x -> "- `" * String(x[1]) * "`: " * string(x[2]), CEnum.name_value_pairs(Lib.DSSSaveFlags)), "\n") )
+"""
+Lib.DSSSaveFlags